+++ apps/agent/.capsule/index.ts
@@ -1,4 +1,3 @@
-// @ts-ignore
 import { createRoot } from '@capsuletech/web-core/create';
 import { Bootstrap } from './bootstrap';